Add basic method to get water by period

+++ src/tasks/WaterLevelTask.h
@@ -22,7 +22,7 @@ private:
 
   const float W1 = 341;
   const float W2 = 682;
-  const unsigned int PEA = 60000;
+  const unsigned long int PEA = 60000;
 
   unsigned long int elapsedTime
   //Lcd* lcd;
